% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG.R
\name{build_model_set}
\alias{build_model_set}
\title{Build a model set.}
\usage{
build_model_set(..., .common)
}
\arguments{
\item{...}{Named arugments, which each are a lists of formulas defining the
paths of a causal model.}

\item{.common}{A list of formulas that contain causal paths that are common
to each model.}
}
\value{
A list of models, each of class \code{matrix} and \code{DAG}.
}
\description{
This is a convenience function to quickly and clearly define a set of causal
models. Supply a list of formulas for each model, using either `list()`,
or `c()`. Formulas should be of the form `parent ~ child` and describe each
path in your model. Multiple children of a single parent can be combined
into a single formula: `parent ~ child1 + child2`.
}
\section{Note This function uses `ggm}{
:DAG`
}

\examples{
(m <- build_model_set(
  A = c(a~b, b~c),
  B = c(b~a, c~b),
  .common = c(d~a)))
plot_model_set(m)
}
