% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cichlids}
\alias{cichlids}
\title{Cichlid traits and the evolution of cooperative breeding.}
\format{
An object of class \code{data.frame} with 69 rows and 5 columns.
}
\source{
Dey, C.J., O'Connor, C.M., Wilkinson, H., Shultz, S., Balshine, S. &
Fitzpatrick, J.L. 2017. Direct benefits and evolutionary transitions to
complex societies. Nat Ecol Evol 1: 137.
}
\usage{
cichlids
}
\description{
A data set with binary traits, used in an analysis on the evolution of
cooperative breeding by Dey et al 2017. Variable names are shortened for
easy of use and consist of cooperative breeding (C), mating system (M),
parental care (P), social grouping (G) and diet (D). All traits are coded as
two level factors.
}
\keyword{datasets}
