% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cichlids_tree}
\alias{cichlids_tree}
\title{Cichlid phylogeny.}
\format{
An object of class \code{phylo} of length 4.
}
\source{
Dey, C.J., O'Connor, C.M., Wilkinson, H., Shultz, S., Balshine, S. &
Fitzpatrick, J.L. 2017. Direct benefits and evolutionary transitions to
complex societies. Nat Ecol Evol 1: 137.
}
\usage{
cichlids_tree
}
\description{
The phylogenetic tree of cichlid species that accompanies the \code{cichlids}
dataset. The phylogeny is based on five nuclear genes and three
mitochondrial genes.
}
\keyword{datasets}
