% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG.R
\name{est_DAG_binary}
\alias{est_DAG_binary}
\title{Add standardized path coefficients to a binary DAG.}
\usage{
est_DAG_binary(DAG, data, tree)
}
\arguments{
\item{DAG}{A directed acyclic graph, typically created with \code{DAG}.}

\item{data}{A \code{data.frame} with binary data. Variables can be a numberic vector (with all
values either 0 or 1), a factor (with two levels) or character vector (with two unique values).}

\item{tree}{A phylogenetic tree of class \code{pylo}.}
}
\value{
An object of class \code{binary_fitted_DAG}.
}
\description{
Add standardized path coefficients to a binary DAG.
}
