% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_plot.R
\name{plot.phylopath_summary}
\alias{plot.phylopath_summary}
\title{Plot the comparison of a set of causal models.}
\usage{
\method{plot}{phylopath_summary}(x, cut_off = 2, ...)
}
\arguments{
\item{x}{A \code{phylopath_summary} object, obtained by calling \code{summary()} on the result of
\code{\link[=phylo_path]{phylo_path()}}.}

\item{cut_off}{The CICc difference within which models are highlighted as having support
comparable to the best model.}

\item{...}{Not used.}
}
\value{
A \code{ggplot} object.
}
\description{
Shows the relative support for each causal model as a bar chart of model weights, ordered from
best to worst. Models within \code{cut_off} CICc of the best model are highlighted, since those are
usually considered to have comparable support. Each bar is labelled with the p-value of the
d-separation test, where a significant value means the model is rejected by the data.
}
\examples{
  candidates <- define_model_set(
    A = NL ~ BM,
    B = NL ~ LS,
    .common = c(LS ~ BM, DD ~ NL)
  )
  p <- phylo_path(candidates, rhino, rhino_tree)
  plot(summary(p))
}
