% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_plot.R
\name{plot_model_set}
\alias{plot_model_set}
\title{Plot several causal hypothesis at once.}
\usage{
plot_model_set(models, algorithm = "kk", text_size = 5, box_x = 12,
  box_y = 10, edge_width = 1, curvature = 0.05, arrow = grid::arrow(type
  = "closed", 15, grid::unit(10, "points")))
}
\arguments{
\item{models}{A list of \code{DAG} objects.}

\item{algorithm}{A layout algorithm from \code{igraph}, see \code{\link[ggraph]{create_layout}}.
By default, uses the Kamada-Kawai layout algorithm. Another good option is \code{"sugiyama"},
which is designed to minimize overlapping edges in DAGs. However, it can often plot nodes too
close together.}

\item{text_size}{Size of the node label text.}

\item{box_x}{To avoid the arrows colliding with the nodes, specify the rectangular dimensions of
an invisible box around each node. If you have long labels, you need to increase this.}

\item{box_y}{To avoid the arrows colliding with the nodes, specify the rectangular dimensions of
an invisible box around each node. If you have multi-line labels, you need to increase this.}

\item{edge_width}{Width of the edges.}

\item{curvature}{Curvature of the edges. A slight curvature can look pretty.}

\item{arrow}{A \code{grid::arrow} object, specifying the shape and size of the arrowheads.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot several causal hypothesis at once.
}
\examples{
m <- list(one = DAG(a ~ b + c + d), two = DAG(a ~ b, b ~ c, d ~ d))
plot_model_set(m)
plot_model_set(m, "sugiyama")
}
