% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{red_list}
\alias{red_list}
\title{Data on brain size, life history and vulnerability to extinction}
\format{An object of class \code{data.frame} with 474 rows and 7 columns.}
\source{
Gonzalez-Voyer A, Gonzalez-Suarez M, Vila C, Revilla E (2016) Larger
brain size indirectly increases vulnerability to extinction in mammals.
Evolution 70:1364-1375. doi: 10.1111/evo.12943
}
\usage{
red_list
}
\description{
A dataset with continuous variables affecting the conservation \code{Status} of
mammalian species (the IUCN red list of threatened species).
}
\details{
It includes the following variables: brain size (\code{Br}), body size (\code{B}),
gestation period (\code{G}), litter size (\code{L}), weening age (\code{W}), population
density (\code{P}) and vulnerability to extinction (\code{Status}).
}
\keyword{datasets}
