% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{red_list_tree}
\alias{red_list_tree}
\title{Mammalian phylogeny}
\format{An object of class \code{phylo} of length 4.}
\source{
Gonzalez-Voyer, A. Gonzalez-Suarez M. Vila C. and Revilla E. 2016.
Larger brain size indirectly increases vulnerability to extinction in
mammals. Evolution 70:1364-1375. doi: 10.1111/evo.12943.

Bininda-Emonds, O. R. P., M. Cardillo, K. E. Jones, R. D. E. MacPhee, R. M. D.
Beck, R. Grenyer, S. A. Price, R. A. Vos, J. L. Gittleman, and A. Purvis.
2007. The delayed rise of present-day mammals. Nature 446:507-512.

Fritz, S. A., O. R. P. Bininda-Emonds, and A. Purvis. 2009. Geographical
variation in predictors of mammalian extinction risk: big is bad, but
only in the tropics. Ecol. Lett. 12:538-549.
}
\usage{
red_list_tree
}
\description{
This is the accompanying phylogeny for the red_list data set. It is based
on the updated mammalian supertree by Bininda-Emonds et al. 2007 & Fritz
et al. 2009.
}
\keyword{datasets}
