% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{guide_train.edge_colourbar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggraph}{\code{\link[ggraph]{guide_train.edge_colourbar}}}
}}

