% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{rhino}
\alias{rhino}
\title{Rhinogrades traits.}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\source{
Gonzalez-Voyer A & von Hardenberg A. 2014. An Introduction to
Phylogenetic Path Analysis. Chapter 8. In: Garamszegi LZ (ed.), Modern
Phylogenetic Comparative Methods and Their Application in Evolutionary
Biology. pp. 201-229. Springer-Verlag Berlin Heidelberg.
doi:10.1111/j.1558-5646.2012.01790.x
}
\usage{
rhino
}
\description{
A simulated dataset, as used by Gonzalez-Voyer and Von Hardenberg as an
example, containing variables on body mass (BM), litter size (LS), nose
length (NL), dispersal distance (DD) and range size (RS).
}
\keyword{datasets}
