% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{rhino_tree}
\alias{rhino_tree}
\title{Rhinogrades phylogeny.}
\format{
An object of class \code{phylo} of length 4.
}
\source{
Gonzalez-Voyer A & von Hardenberg A. 2014. An Introduction to
Phylogenetic Path Analysis. Chapter 8. In: Garamszegi LZ (ed.), Modern
Phylogenetic Comparative Methods and Their Application in Evolutionary
Biology. pp. 201-229. Springer-Verlag Berlin Heidelberg.
doi:10.1111/j.1558-5646.2012.01790.x
}
\usage{
rhino_tree
}
\description{
A phylogenetic tree for the 100 species of the \code{rhino} dataset.
}
\keyword{datasets}
