% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_plot.R
\name{show_warnings}
\alias{show_warnings}
\title{Print out warnings from a phylopath analysis.}
\usage{
show_warnings(phylopath = NULL)
}
\arguments{
\item{phylopath}{A \code{phylopath} object of which the warnings should be printed.}
}
\description{
Use this function after running \code{phylo_path()} to conveniently print any generated warnings
to the screen. You can either provide no arguments, which will only work if you run it directly
after the analysis, or you have to provide the \code{phylopath} object manually.
}
