% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{average}
\alias{average}
\title{Extract and average the best supported models for a phylogenetic path
analysis.}
\usage{
average(phylopath, cut_off = 2, method = "conditional", ...)
}
\arguments{
\item{phylopath}{An object of class \code{phylopath}.}

\item{cut_off}{The CICc cut-off used to select the best models. Use
\code{Inf} to average over all models. Use the \code{best} function to
only use the top model.}

\item{method}{Either \code{"full"} or \code{"conditional"}. The methods
differ in how they deal with averaging a path coefficient where the path is
absent in some of the models. The full method sets the coefficent (and the
variance) for the missing paths to zero, meaning paths that are missing in
some models will shrink towards zero. The conditional method only averages
over models where the path appears, making it more sensitive to small
effects. Following von Hardenberg & Gonzalez-Voyer 2013, conditional
averaging is set as the default. Also see \link[MuMIn]{model.avg}.}

\item{...}{Additional arguments passed to \link[MuMIn]{par.avg}.

  For details on the error calculations, see \link[MuMIn]{par.avg}.}
}
\value{
An object of class \code{fitted_DAG}.
}
\description{
Extract and average the best supported models for a phylogenetic path
analysis.
}

