% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG.R
\name{average_DAGs}
\alias{average_DAGs}
\title{Perform model averaging on a list of DAGs.}
\usage{
average_DAGs(coef, std_error = NULL, weights = rep(1, length(coef)),
  method = "conditional", ...)
}
\arguments{
\item{coef}{A list of estimated DAGs, usually created with \code{est_DAG}.}

\item{std_error}{Optionally, a list of associated matrices standard errors,
one matrix for each DAG. Can also be obtained from \code{est_DAG}.}

\item{weights}{A vector of associated model weights.}

\item{method}{Either \code{"full"} or \code{"conditional"}. The methods
differ in how they deal with averaging a path coefficient where the path is
absent in some of the models. The full method sets the coefficent (and the
variance) for the missing paths to zero, meaning paths that are missing in
some models will shrink towards zero. The conditional method only averages
over models where the path appears, making it more sensitive to small
effects. Following von Hardenberg & Gonzalez-Voyer 2013, conditional
averaging is set as the default. Also see \link[MuMIn]{model.avg}.}

\item{...}{Additional arguments passed to \link[MuMIn]{par.avg}.

  For details on the error calculations, see \link[MuMIn]{par.avg}.}
}
\value{
An object of class \code{fitted_DAG}, including standard errors and
  confidence inverals if \code{std_error} was supplied.
}
\description{
Perform model averaging on a list of DAGs.
}

