% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{best}
\alias{best}
\title{Extract and estimate the best supported model for a phylogenetic path
analysis.}
\usage{
best(phylopath)
}
\arguments{
\item{phylopath}{An object of class \code{phylopath}.}
}
\value{
An object of class \code{fitted_DAG}.
}
\description{
Extract and estimate the best supported model for a phylogenetic path
analysis.
}

