% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{phylo_path}
\alias{phylo_path}
\title{Compare causal models in a phylogenetic context.}
\usage{
phylo_path(models, data, tree, cor_fun = ape::corPagel, cut_off = 2)
}
\arguments{
\item{models}{A list of directed acyclic graphs. These are matrices,
typically created with \code{DAG}.}

\item{data}{A \code{data.frame} with data.}

\item{tree}{A phylogenetic tree of class \code{pylo}.}

\item{cor_fun}{A function that creates a \code{corStruct} object, typically
one of the cor function from the \code{ape}, such as \code{corBrownian},
\code{corPagel} etc.}

\item{cut_off}{The cut off value to decide which models to include in the
average model, in terms of delta CICc.}
}
\value{
A table with relevant statistics for each model, including the C
  statistic, the associated p-values, information criterions, and model
  weigths.
}
\description{
Compare causal models in a phylogenetic context.
}

