% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_and_plot.R
\name{plot_model_set}
\alias{plot_model_set}
\title{Plot several causal hypothesis at once.}
\usage{
plot_model_set(
  model_set,
  labels = NULL,
  algorithm = "kk",
  manual_layout = NULL,
  text_size = 5,
  box_x = 12,
  box_y = 10,
  edge_width = 1,
  curvature = 0.05,
  rotation = 0,
  flip_x = FALSE,
  flip_y = FALSE,
  nrow = NULL,
  arrow = grid::arrow(type = "closed", 15, grid::unit(10, "points"))
)
}
\arguments{
\item{model_set}{A list of \code{DAG} objects, usually created with \code{\link[=define_model_set]{define_model_set()}}.}

\item{labels}{An optional set of labels to use for the nodes. This should be a named vector, of
the form \code{c(var1 = "label1", var2 = "label2")}.
If left at `NULL``, the variable names of the DAGs are used.}

\item{algorithm}{A layout algorithm from \code{igraph}, see
\code{\link[ggraph:create_layout]{ggraph::create_layout()}}. By default, uses the Kamada-Kawai
layout algorithm. Another good option is \code{"sugiyama"}, which is
designed to minimize edge crossing in DAGs. However, it can often plot
nodes too close together.}

\item{manual_layout}{Alternatively, precisely define the layout yourself, by providing a
\code{data.frame} that at least has a column \code{name} with all variable names, and columns \code{x} and \code{y}
with positions to be plotted. Setting this parameter overrides \code{algorithm} but other changes,
such as \code{rotation} and \code{flip}s will still be applied.}

\item{text_size}{Size of the node label text.}

\item{box_x}{To avoid the arrows colliding with the nodes, specify the
rectangular dimensions of an invisible box around each node. If you have
long labels, you need to increase this.}

\item{box_y}{To avoid the arrows colliding with the nodes, specify the
rectangular dimensions of an invisible box around each node. If you have
multi-line labels, you need to increase this.}

\item{edge_width}{Width of the edges.}

\item{curvature}{Curvature of the edges. A slight curvature can look pretty.}

\item{rotation}{Supply the degrees you want to rotate the layout by. This is useful in order to
put rotate your upstream nodes towards the top if needed.}

\item{flip_x}{Whether to flip the node positions horizontally.}

\item{flip_y}{Whether to flip the node positions vertically.}

\item{nrow}{Number of rows to display the models on.}

\item{arrow}{A \code{grid::arrow} object, specifying the shape and size of the arrowheads.

The order of facets is taken from the ordering of the list, with the facet
labels coming from the names of the list. If the list is unnamed, sequential
lettering is used.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot several causal hypothesis at once.
}
\examples{
m <- list(one = DAG(a ~ b + c + d), two = DAG(a ~ b, b ~ c, d ~ d))
plot_model_set(m)
plot_model_set(m, algorithm = "sugiyama")
}
